requirements.txt
httpx==0.23.0
playwright==1.24.1
python-dotenv==0.20.0
pytz==2022.1
rich==12.5.1
typer==0.6.1
